refactor: Change code such that they're broken into modules
Was mainly moving files.  Only change was in "Phase" where i made the member d_gameEngine protected.
