+++ src/main/java/ca/concordia/soen6441/project/GameEngine.java
@@ -1,6 +1,10 @@
 package ca.concordia.soen6441.project;
 
+import ca.concordia.soen6441.project.gameplay.PlayerImpl;
 import ca.concordia.soen6441.project.interfaces.*;
+import ca.concordia.soen6441.project.map.*;
+import ca.concordia.soen6441.project.phases.Phase;
+import ca.concordia.soen6441.project.phases.PreLoad;
 
 import java.io.FileNotFoundException;
 import java.util.stream.Collectors;
