+++ src/main/java/ca/concordia/soen6441/project/map/ContinentImpl.java
@@ -1,4 +1,4 @@
-package ca.concordia.soen6441.project;
+package ca.concordia.soen6441.project.map;
 
 import ca.concordia.soen6441.project.interfaces.Continent;
 import ca.concordia.soen6441.project.interfaces.MapComponent;
