+++ src/main/java/ca/concordia/soen6441/project/phases/AssignReinforcements.java
@@ -1,10 +1,10 @@
-package ca.concordia.soen6441.project;
+package ca.concordia.soen6441.project.phases;
 
+import ca.concordia.soen6441.project.GameEngine;
 import ca.concordia.soen6441.project.interfaces.Player;
 import ca.concordia.soen6441.project.interfaces.Continent;
 import ca.concordia.soen6441.project.interfaces.Country;
 
-import java.util.ArrayList;
 import java.util.Map;
 
 public class AssignReinforcements extends MainPlay {
