+++ src/main/java/ca/concordia/soen6441/project/phases/Phase.java
@@ -1,9 +1,9 @@
-package ca.concordia.soen6441.project;
+package ca.concordia.soen6441.project.phases;
 
-import ca.concordia.soen6441.project.interfaces.Player;
+import ca.concordia.soen6441.project.GameEngine;
 
 public abstract class Phase {
-    GameEngine d_gameEngine;
+    protected GameEngine d_gameEngine;
 
     public Phase(GameEngine p_gameEngine) {
         this.d_gameEngine = p_gameEngine;
