+++ src/main/java/ca/concordia/soen6441/project/phases/Startup.java
@@ -1,6 +1,8 @@
-package ca.concordia.soen6441.project;
+package ca.concordia.soen6441.project.phases;
 
-import ca.concordia.soen6441.project.interfaces.Player;
+import ca.concordia.soen6441.project.GameEngine;
+import ca.concordia.soen6441.project.map.InvalidMapFileException;
+import ca.concordia.soen6441.project.gameplay.CountryAssignment;
 
 import java.io.FileNotFoundException;
