+++ src/test/java/ca/concordia/soen6441/project/phases/IssueOrderTest.java
@@ -1,7 +1,9 @@
-package ca.concordia.soen6441.project;
+package ca.concordia.soen6441.project.phases;
 
 import static org.mockito.Mockito.*;
 
+import ca.concordia.soen6441.project.GameEngine;
+import ca.concordia.soen6441.project.gameplay.orders.Deploy;
 import ca.concordia.soen6441.project.interfaces.Country;
 import ca.concordia.soen6441.project.interfaces.Player;
 import org.junit.jupiter.api.BeforeEach;
